+++ example/lib/device_scan_screen.dart
@@ -2,6 +2,7 @@ import 'package:chafon_h103_rfid/chafon_h103_rfid.dart';
 import 'package:flutter/material.dart';
 import 'package:permission_handler/permission_handler.dart';
 import 'BluetoothDeviceModel.dart';
+import 'functions.dart';
 
 class DeviceScanScreen extends StatefulWidget {
   const DeviceScanScreen({super.key});
@@ -25,15 +26,25 @@ class _DeviceScanScreenState extends State<DeviceScanScreen> {
   void initState() {
     super.initState();
     checkPermissions();
+
+    ChafonH103RfidService.isConnected().then((connected) {
+      if (connected == true && mounted) {
+        Navigator.pushReplacement(
+          context,
+          MaterialPageRoute(builder: (_) => const Functions()),
+        );
+      }
+    });
+
     ChafonH103RfidService.initCallbacks(
       onDeviceFound: (device) {
         final newDevice = BluetoothDeviceModel(
-          name: device['name'] ?? 'Naməlum',
+          name: device['name'] ?? 'Unknown',
           address: device['address']!,
           rssi: device['rssi'] ?? 0,
         );
 
-        // Əgər bu cihaz artıq siyahıda yoxdursa əlavə et
+
         if (!devices.any((d) => d.address == newDevice.address)) {
           setState(() {
             devices.add(newDevice);
@@ -80,7 +91,7 @@ class _DeviceScanScreenState extends State<DeviceScanScreen> {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      appBar: AppBar(title: const Text('Bluetooth Cihazları')),
+      appBar: AppBar(title: const Text('Bluetooth Device')),
       body: Column(
         children: [
           if (isLoading) const LinearProgressIndicator(),
@@ -98,7 +109,7 @@ class _DeviceScanScreenState extends State<DeviceScanScreen> {
                       ? const Icon(Icons.check_circle, color: Colors.green)
                       : ElevatedButton(
                           onPressed: () => _connectToDevice(device),
-                          child: const Text("Qoşul"),
+                          child: const Text("Connect"),
                         ),
                 );
               },
@@ -109,10 +120,10 @@ class _DeviceScanScreenState extends State<DeviceScanScreen> {
               padding: const EdgeInsets.all(16.0),
               child: ElevatedButton.icon(
                 onPressed: () {
-                  Navigator.pushNamed(context, "/functions");
+                  Navigator.push(context, MaterialPageRoute(builder: (context)=>Functions()));
                 },
                 icon: const Icon(Icons.play_circle),
-                label: const Text("Funksiyalara keç"),
+                label: const Text("Functions"),
               ),
             ),
         ],