+++ example/lib/main.dart
@@ -1,14 +1,10 @@
-import 'package:chafon_h103_rfid_example/presentaion/login/login_page.dart';
 import 'package:flutter/material.dart';
+import 'device_scan_screen.dart';
 
 void main() {
   runApp(
     MaterialApp(
-      home: const LoginScreen(),
-      routes: {
-        "/functions": (context) =>
-            LoginScreen(),
-      },
+      home: const DeviceScanScreen(),
     ),
   );
 }