+++ CHANGELOG.md
@@ -1,3 +1,8 @@
 ## 0.0.1
 
 * TODO: Describe initial release.
+
+## 0.1.0
+- Added disconnect button in Settings tab.
+- Added auto-navigation to Functions screen if device already connected.
+- Translated UI text to English for better understanding.